Add User Todos Laravel API: now each user access to his todos

+++ app/Http/Controllers/TodosController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Todo;
 use Illuminate\Http\Request;
+use function foo\func;
 
 class TodosController extends Controller {
     /**
@@ -12,7 +13,8 @@ class TodosController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function index() {
-        return Todo::all();
+        return Todo::where('user_id', auth()->user()->id)->get();
+
     }
 
     /**
@@ -27,7 +29,11 @@ class TodosController extends Controller {
             'completed' => 'required|boolean',
         ]);
 
-        $todo = Todo::create($data);
+        $todo = Todo::create([
+            'user_id' => auth()->user()->id,
+            'title' => $request->title,
+            'completed' => $request->completed,
+        ]);
 
         return response($todo, 201);
     }
@@ -40,6 +46,9 @@ class TodosController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Todo $todo) {
+        if ($todo->user_id != auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
         $data = $request->validate([
             'title' => 'required|string',
             'completed' => 'required|boolean',
@@ -55,7 +64,7 @@ class TodosController extends Controller {
             'completed' => 'required|boolean',
         ]);
 
-        Todo::query()->update($data);
+        Todo::where('user_id', auth()->user()->id)->update($data);
         return response('Updated', 200);
     }
 
@@ -67,11 +76,29 @@ class TodosController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function destroy(Todo $todo) {
+        if ($todo->user_id != auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
+
         $todo->delete();
         return response('Deleted todo item', 200);
     }
 
     public function destroyCompleted(Request $request) {
+
+        $todosToDelete = $request->todos;
+        $userTodoId = auth()->user()->todos->map(function ($todo) {
+            return $todo->id;
+        });
+
+        $valid = collect($todosToDelete)->every(function ($value, $key) use ($userTodoId) {
+            return $userTodoId->contains($value);
+        });
+
+        if (!$valid) {
+            return response()->json('Unauthorized', 401);
+        }
+
         $request->validate([
             'todos' => 'required|array',
         ]);
